jebnf-core/src/test/java/de/weltraumschaf/jebnf/ast/nodes/RuleNodeTest.java
/*
 * LICENSE
 *
 * "THE BEER-WARE LICENSE" (Revision 43):
 * "Sven Strittmatter" <weltraumschaf(at)googlemail(dot)com> wrote this file.
 * As long as you retain this notice you can do whatever you want with
 * this stuff. If we meet some day, and you think this stuff is worth it,
 * you can buy me a non alcohol free beer in return.
 *
 * Copyright (C) 2012 "Sven Strittmatter" <weltraumschaf(at)googlemail(dot)com>
 */
package de.weltraumschaf.jebnf.ast.nodes;

import de.weltraumschaf.jebnf.ast.Node;
import de.weltraumschaf.jebnf.ast.NodeType;
import static org.junit.Assert.assertEquals;
import org.junit.Test;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

/**
 *
 * @author Sven Strittmatter <weltraumschaf@googlemail.com>
 */
public class RuleNodeTest {

    @Test public void testToString() {
        final RuleNode rule = RuleNode.newInstance("foo");
        assertEquals("<RULE NAME=foo>", rule.toString());
        final Node node1 = mock(Node.class);
        when(node1.toString()).thenReturn("<foo>");
        rule.addChild(node1);
        final Node node2 = mock(Node.class);
        when(node2.toString()).thenReturn("<bar>");
        rule.addChild(node2);
        assertEquals("<RULE NAME=foo>\n"
                   + "<foo>\n"
                   + "<bar>", rule.toString());
    }

    @Test public void getType() {
        assertEquals(NodeType.RULE, RuleNode.newInstance().getType());
    }

}
